Android: Remove the designSupport from the actual library.
I hit a merge conflict problem downstream with another app/library that was trying to specify the theming. This seemed to create a merge conflict.
I think this best strategy is to never declare the design/theme in the actual library, and only do it in the final application.
So I removed it from the iup library, but left it in the test apps. This seems to work. (It looks like I never added it to the other iup* libraries.)