Mac: Crashing on 'exit after last window is closed' is happening consistently now.
I think Apple has changed something in recent versions of macOS that now triggers this all the time.
Ultimately, the problem seems to be that there isn't a time in the Iup API late enough for us to call [NSAutoreleasePool drain]. In the current close path, we are still inside a Cocoa callback shutting things down, so it is too early to call drain, which then seems to lead Cocoa to crash since it is still depending on the pool.

I think the only solution is to not call drain, and let this technically leak. In practice, this is not actually a leak since the app is supposed to be quiting, and thus you don't need to clean up memory since the OS will clean up. The only place this might be a problem is people trying to force the old Iup paradigm on IupNext, but as I have documented, that is never going to work in the end.

I did experiment with dispatch_async, but the problem was that it never gets called, presumably since we don't have an event loop any more and never a chance for it to run.