Mac: Label: New attribute SELECTABLE (Mac only): Allows the label text to be selectable with the mouse for copy to clipboard.
This feature originally had a bad bug in older versions of macOS, but in testing with the latest macOS, it appears to be fixed. However, older macOS will still have problems. I don't know which version this was fixed in. By default, it is off, so it won't be a problem unless the user turns it on. They can decide which versions of macOS they want to support or conditionally enable it based on OS version.